+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version 0.2
 * Fixed an issue that caused bad sectors not to be reported in the stats file.
+* Program now pauses when a device is removed during the stress test, and resumes automatically once the device is reconnected.
 
 # Version 0.1
 Initial release